Updates parser and type checking to properly handle code contracts
We move the contracts specification from the function body to the
function signature. We also want to be able to handle contracts when
verifying multiple source files (e.g., contracts specified in .h files).
Finally, we must prepare the parser and the type-checking engine
to handle new assigns clause constructs, which enables support
for arrays and structs.

Signed-off-by: Felipe R. Monteiro <felisous@amazon.com>